02_neurosan_multi_agent_orchestration/requirements.txt
neuro-san>=0.6.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
azure-identity>=1.16.0
python-dotenv>=1.0.0
requests>=2.32.0
pyhocon>=0.3.61
